+++ src/index.ts
@@ -1,5 +1,8 @@
+const COLON = 58
+const WILDCARD = 42
+
 interface RadixNode<T> {
-    children: Map<string, RadixNode<T>>
+    children: Map<string | number, RadixNode<T>>
     static?: Record<string, T>
     store?: T
     name?: string
@@ -38,13 +41,11 @@ export class Raikiri<T> {
             paths[paths.length - 1] += `${part}/`
         })
 
-        if (!path.endsWith('/')) {
-            // Remove trailing slash
-            if (path !== '/')
-                paths[paths.length - 1] = paths[paths.length - 1].slice(
-                    0,
-                    paths[paths.length - 1].length - 1
-                )
+        if (!path.endsWith('/') && path !== '/') {
+            paths[paths.length - 1] = paths[paths.length - 1].slice(
+                0,
+                paths[paths.length - 1].length - 1
+            )
         }
 
         if (paths.length > 1 && paths[paths.length - 1] === '') paths.pop()
@@ -61,7 +62,7 @@ export class Raikiri<T> {
         if (paths[1] === '*' && paths[0] === '/' && paths.length === 2) {
             const wildcardNode = createNode<T>()
             wildcardNode.store = store
-            this.root[method].children.set('*', wildcardNode)
+            this.root[method].children.set(WILDCARD, wildcardNode)
 
             return
         }
@@ -71,33 +72,53 @@ export class Raikiri<T> {
             const isLast = i === paths.length - 1
             let iterated = false
 
-            // console.log(path)
-
-            // If done correctly, there should be only 1 static key
             for (const key of node.children.keys()) {
-                let prefix = ''
+                let prefix: string | number = ''
 
-                for (const [charIndex, charKey] of key.split('').entries()) {
-                    if (path[charIndex] === charKey) prefix += charKey
-                    else break
-                }
+                if (typeof key === 'string')
+                    for (const [charIndex, charKey] of key
+                        .split('')
+                        .entries()) {
+                        if (path[charIndex] === charKey) prefix += charKey
+                        else break
+                    }
+                else prefix = 58
 
                 if (!prefix) {
                     iterated = true
                     break
                 }
 
+                const prefixLength =
+                    typeof prefix === 'number' ? 1 : prefix.length
+
                 if (node.children.has(prefix)) {
                     iterated = true
                     node = node.children.get(prefix)!
 
-                    const fracture = path.slice(prefix.length)
+                    const fracture = path.slice(prefixLength)
 
-                    if (prefix !== path && !prefix.startsWith(':')) {
-                        const part = path.slice(prefix.length)
+                    if (prefix !== path && prefix !== COLON) {
+                        const part = path.slice(prefixLength)
 
                         if (!node.children.has(fracture))
-                            node.children.set(part, createNode<T>())
+                            if (
+                                node.children.has(COLON) ||
+                                node.children.has(WILDCARD)
+                            ) {
+                                // Move static above colon and wildcard
+                                const ordered = new Map()
+
+                                ordered.set(part, createNode<T>())
+
+                                for (const [
+                                    key,
+                                    value
+                                ] of node.children.entries())
+                                    ordered.set(key, value)
+
+                                node.children = ordered
+                            } else node.children.set(part, createNode<T>())
 
                         if (!node.static) node.static = {}
                         node.static[part] = store
@@ -111,25 +132,27 @@ export class Raikiri<T> {
                 if (prefix && prefix !== '/' && prefix !== path) {
                     iterated = true
 
-                    let branchNode = createNode<T>()
-                    const migrate = key.slice(prefix.length)
-                    branchNode.children.set(migrate, {
-                        ...node.children.get(key)!
+                    // Newly created path: is empty
+                    const branch = path.slice(prefixLength)
+                    const migrate =
+                        typeof key === 'number' ? key : key.slice(prefixLength)
+
+                    const branchNode = createNode<T>({
+                        [migrate]: node.children.get(key)!,
+                        [branch]: createNode()
                     })
+
                     node.children.delete(key)
 
-                    if (!migrate.includes(':')) {
+                    if (migrate !== COLON) {
                         if (!node.static) node.static = {}
                         node.static[migrate] =
                             branchNode.children.get(migrate)?.store!
                     }
 
-                    // Newly created path: is empty
-                    const branch = path.slice(prefix.length)
-                    branchNode.children.set(branch, createNode<T>())
                     node.children.set(prefix, branchNode)
 
-                    if (isLast && !prefix.startsWith(':')) {
+                    if (isLast && prefix !== COLON) {
                         if (!node.static) node.static = {}
                         node.static[path] = store
                     }
@@ -141,11 +164,11 @@ export class Raikiri<T> {
             if (path.startsWith(':')) {
                 const paramNode = createNode<T>()
                 paramNode.name = path.slice(1)
-                node.children.set(':', paramNode)
+                node.children.set(COLON, paramNode)
                 node = paramNode
             } else if (isLast && path === '*') {
-                node.children.set('*', createNode())
-                node = node.children.get(path)!
+                node.children.set(WILDCARD, createNode())
+                node = node.children.get(WILDCARD)!
             } else {
                 if (isLast) {
                     if (!node.static) node.static = {}
@@ -159,8 +182,6 @@ export class Raikiri<T> {
         }
 
         node.store = store
-
-        // console.log(this.root[method])
     }
 
     match(method: string, path: string) {
@@ -174,85 +195,83 @@ export class Raikiri<T> {
                 params: {}
             }
 
-        // console.log(path, node.children.has('*'))
-
-        // if (
-        //     node.children.has('*') ||
-        //     node.children.get('/')?.children.has('*')
-        // ) {
-        //     return {
-        //         store:
-        //             node.children.get('*')?.store ||
-        //             node.children.get('/')?.children.get('*')?.store,
-        //         params: {
-        //             '*': path
-        //         }
-        //     }
-        // }
-
+        const clear = path.length - 2
         let params: { [key: string]: string } = {}
         let depth = 0
 
-        while (depth <= path.length - 1) {
+        while (true) {
             let isEnd = true
             const children = node.children
 
             for (const fracture of children.keys()) {
-                const length = depth + fracture.length
-                const part = path.slice(depth, length)
+                if (typeof fracture === 'number') {
+                    if (fracture === COLON) {
+                        node = children.get(COLON)!
+
+                        const index = path.indexOf('/', depth + 1)
+                        const value =
+                            index === -1
+                                ? path.slice(depth)
+                                : path.slice(depth, index)
+
+                        params[node.name!] = value
+
+                        depth += value.length
+                        isEnd = false
+                        break
+                    }
+                    // Since it's special characters and not : then it's wildcard
+                    else {
+                        params['*'] = path.slice(depth)
+
+                        return {
+                            store: node.children.get(WILDCARD)!.store,
+                            params
+                        }
+                    }
+
+                    // Special characters should all be matched above, abort if not
+                    break
+                }
+
+                const current = depth + fracture.length
+                const part = path.slice(depth, current)
 
                 const root = node.static?.[part]
-                if (root && length > path.length - 2) {
+                if (root) {
+                    /**
+                     * Suppose there are:
+                     * - /id/:id/name/:name
+                     * - /id/:id/name/a
+                     *
+                     * Then if you pass in
+                     * - /id/1/name/ame
+                     *
+                     * Since fracture is /a, path is /ame then part become /a
+                     * It will be matched in node.static[part]
+                     *
+                     * This means we need to check path left by using current <= clear
+                     */
+                    if (current <= clear) continue
+
                     return {
                         store: root,
                         params
                     }
                 }
 
-                if (children.has(part) && fracture !== ':') {
+                if (children.has(part)) {
                     node = children.get(part)!
                     depth += part.length
                     isEnd = false
                     break
-                }
-
-                if (children.has(':')) {
-                    const index = path.indexOf('/', depth + 1)
-                    const value =
-                        index === -1
-                            ? path.slice(depth)
-                            : path.slice(depth, index)
-
-                    node = children.get(':')!
-                    params[node.name!] = value
-
-                    // console.log('dyn', node.name, value)
-
-                    depth += value.length
-                    isEnd = false
-                    break
-                }
-
-                if (children.has('*')) {
-                    params['*'] = path.slice(depth)
-
-                    return {
-                        store: node.children.get('*')!.store,
-                        params
-                    }
                 }
             }
 
             if (isEnd) break
         }
 
-        if (depth > path.length - 2) {
-            const store = node.store
-
-            if (store) {
-                return { store: node.store, params }
-            }
-        }
+        if (node.store) return { store: node.store, params }
     }
 }